+++ copy_with_extension_gen/lib/src/field_info.dart
@@ -1,8 +1,10 @@
 import 'package:analyzer/dart/constant/value.dart' show DartObject;
-import 'package:analyzer/dart/element/element2.dart'
-    show ClassElement2, FieldElement2, FormalParameterElement;
 import 'package:analyzer/dart/element/nullability_suffix.dart';
+import 'package:analyzer/dart/element/type.dart'
+    show ParameterizedType, DartType;
 import 'package:copy_with_extension/copy_with_extension.dart';
+import 'package:copy_with_extension_gen/src/helpers.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:copy_with_extension_gen/src/copy_with_field_annotation.dart';
 import 'package:copy_with_extension_gen/src/helpers.dart'
     show readElementNameOrThrow;
@@ -37,7 +39,7 @@ class ConstructorParameterInfo extends FieldInfo {
         super(
           name: readElementNameOrThrow(element),
           nullable: element.type.nullabilitySuffix != NullabilitySuffix.none,
-          type: element.type.getDisplayString(),
+          type: _fullTypeName(element),
         );
 
   /// Annotation provided by the user with `CopyWithField`.
@@ -70,6 +72,62 @@ class ConstructorParameterInfo extends FieldInfo {
       nullable: field.type.nullabilitySuffix != NullabilitySuffix.none,
       type: field.type.getDisplayString(),
     );
+  }
+
+  /// Returns full type name including namespace for all nested type arguments.
+  static String _fullTypeName(FormalParameterElement element) {
+    final library = element.library2;
+    if (library is! LibraryElement2) {
+      return element.type.getDisplayString();
+    }
+
+    return _typeNameWithPrefix(library, element.type);
+  }
+
+  /// Recursively builds the name for the `type` including import prefixes.
+  static String _typeNameWithPrefix(LibraryElement2 library, DartType type) {
+    final nullability =
+        type.nullabilitySuffix == NullabilitySuffix.question ? '?' : '';
+
+    if (type is ParameterizedType) {
+      final element = type.element3;
+      final name = element != null
+          ? '${_prefixFor(library, element.library2)}${element.name3}'
+          : type.getDisplayString(withNullability: false);
+
+      if (type.typeArguments.isNotEmpty) {
+        final args = type.typeArguments
+            .map((t) => _typeNameWithPrefix(library, t))
+            .join(', ');
+        return '$name<$args>$nullability';
+      } else {
+        return '$name$nullability';
+      }
+    }
+
+    final displayName = type.getDisplayString(withNullability: false);
+    return '${_prefixFor(library, type.element3?.library2)}$displayName$nullability';
+  }
+
+  /// Returns the import prefix for `targetLibrary` if one exists in `library`.
+  static String _prefixFor(
+    LibraryElement2 library,
+    LibraryElement2? targetLibrary,
+  ) {
+    if (targetLibrary == null) return '';
+    final unit = library.fragments.first;
+    for (final PrefixElement2 prefix in unit.prefixes) {
+      for (final LibraryImport import in prefix.imports) {
+        if (import.importedLibrary2 == targetLibrary) {
+          final prefixName = prefix.name3;
+          if (prefixName is String && prefixName.isNotEmpty) {
+            return '$prefixName.';
+          }
+        }
+      }
+    }
+
+    return '';
   }
 
   /// Restores the `CopyWithField` annotation provided by the user.